Kernel compilation unit
